+++ src/main/java/one/util/streamex/DoubleStreamEx.java
@@ -42,6 +42,7 @@ import java.util.function.DoubleToLongFunction;
 import java.util.function.DoubleUnaryOperator;
 import java.util.function.Function;
 import java.util.function.ObjDoubleConsumer;
+import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.stream.Collector;
 import java.util.stream.DoubleStream;
@@ -50,6 +51,7 @@ import java.util.stream.LongStream;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
+import one.util.streamex.StreamExInternals.Box;
 import static one.util.streamex.StreamExInternals.*;
 
 /**
@@ -1927,6 +1929,33 @@ public class DoubleStreamEx extends BaseStreamEx<Double, DoubleStream, Spliterat
     }
 
     /**
+     * Return an ordered stream produced by consecutive calls of the supplied
+     * producer until it returns false.
+     * 
+     * <p>
+     * The producer function may call the passed consumer any number of times
+     * and return true if the producer should be called again or false
+     * otherwise. It's guaranteed that the producer will not be called anymore,
+     * once it returns false.
+     * 
+     * <p>
+     * This method is particularly useful when producer changes the mutable
+     * object which should be left in known state after the full stream
+     * consumption. Note however that if a short-circuiting operation is used,
+     * then the final state of the mutable object cannot be guaranteed.
+     * 
+     * @param producer a predicate which calls the passed consumer to emit
+     *        stream element(s) and returns true if it producer should be
+     *        applied again.
+     * @return the new stream
+     * @since 0.6.0
+     */
+    public static DoubleStreamEx produce(Predicate<DoubleConsumer> producer) {
+        Box<DoubleEmitter> box = new Box<>(null);
+        return (box.a = action -> producer.test(action) ? box.a : null).stream();
+    }
+
+    /**
      * Returns a sequential unordered {@code DoubleStreamEx} of given length
      * which elements are equal to supplied value.
      * 
@@ -1979,14 +2008,20 @@ public class DoubleStreamEx extends BaseStreamEx<Double, DoubleStream, Spliterat
          * nothing more to emit.
          * 
          * <p>
+         * Normally one element is emitted during the {@code next()} method
+         * call. However, it's not restricted: you may emit as many elements as
+         * you want, though in some cases if many elements were emitted they
+         * might be buffered consuming additional memory.
+         * 
+         * <p>
          * It's allowed not to emit anything (don't call the consumer). However
          * if you do this and return new emitter which also does not emit
          * anything, you will end up in endless loop.
          * 
-         * @param cons consumer to be called to emit elements
+         * @param action consumer to be called to emit elements
          * @return next emitter or null
          */
-        DoubleEmitter next(DoubleConsumer cons);
+        DoubleEmitter next(DoubleConsumer action);
 
         /**
          * Returns the spliterator which covers all the elements emitted by this